requirements/base.txt
Django==2.0.7
django_environ==0.4.5
djangorestframework==3.8.2
django_rest_registration==0.3.12
boto3==1.7.67